+++ src/sections/campaign/discover/admin/submissions/v4/shared/feedback-section.jsx
@@ -1,0 +1,42 @@
+import PropTypes from 'prop-types';
+import { Box, Stack } from '@mui/material';
+import FeedbackDisplay from './feedback-display';
+
+export default function FeedbackSection({ submission, isVisible, isClient }) {
+  if (!isVisible || !submission.feedback || submission.feedback.length === 0) {
+    return null;
+  }
+
+  const filteredFeedback = submission.feedback.filter(feedback => {
+    if (isClient) {
+      return feedback.type === 'COMMENT';
+    }
+    return feedback.type;
+  });
+
+  if (filteredFeedback.length === 0) {
+    return null;
+  }
+
+  const latestFeedback = filteredFeedback[0];
+  const showContent = submission.status !== 'CLIENT_FEEDBACK';
+
+  return (
+    <Box sx={{ flex: 1, overflow: 'auto' }}>
+      <Stack spacing={1}>
+        <FeedbackDisplay
+          submission={submission}
+          feedback={latestFeedback}
+          showContent={showContent}
+          isClient={isClient}
+        />
+      </Stack>
+    </Box>
+  );
+}
+
+FeedbackSection.propTypes = {
+  submission: PropTypes.object.isRequired,
+  isVisible: PropTypes.bool.isRequired,
+  isClient: PropTypes.bool.isRequired,
+};
